Make PlayerUIProtocol more extension.
