+++ SoundScape/Model/Audio/PlayerUIProtocol.swift
@@ -1,0 +1,112 @@
+//
+//  PlayerUIProtocol.swift
+//  SoundScape
+//
+//  Created by Astrid on 2021/11/24.
+//
+
+import Foundation
+import UIKit
+
+protocol PlayerUIProtocol: AnyObject {
+    
+    // model
+    
+    //    var nowPlayingURL: URL? { get set }
+    
+    // UI
+    
+    //    var playButtonColor: UIColor { get set } // default
+    
+    var playButton: UIButton { get } // default
+    
+    var caDisplayLink: CADisplayLink? { get set }
+    
+    var playButtonImagePlay: UIImage { get } // default
+    
+    var playButtonImagePause: UIImage { get } // default
+    
+    var playButtonImageStop: UIImage { get } // default
+    
+    var progressView: UIView { get }
+    
+    // method
+    
+    func manipulatePlayer()
+    
+    func updatePlaybackTime(notification: Notification)
+    
+    func updatePlayInfo(notification: Notification)
+    
+    func changeButtonImage() // default
+    
+}
+
+//    default setting
+
+extension PlayerUIProtocol {
+    
+    // model
+    var audioPlayHelper: AudioPlayHelper {
+        return AudioPlayHelper.shared
+    }
+    
+    // UI
+    var playButtonImagePlay: UIImage {
+        let image = UIImage(systemName: CommonUsage.SFSymbol.play) ?? UIImage()
+        return image
+    }
+    
+    var playButtonImagePause: UIImage {
+        let image = UIImage(systemName: CommonUsage.SFSymbol.pause) ?? UIImage()
+        return image
+    }
+    
+    var playButtonImageStop: UIImage {
+        let image = UIImage(systemName: CommonUsage.SFSymbol.stopPlay) ?? UIImage()
+        return image
+    }
+    
+    func manipulatePlayer() {
+        if audioPlayHelper.isPlaying {
+            audioPlayHelper.pause()
+        } else {
+            audioPlayHelper.play()
+        }
+    }
+    
+    func updatePlaybackTime(notification: Notification) {
+        guard let playProgress = notification.userInfo?["UserInfo"] as? PlayProgress else { return }
+        let currentTime = playProgress.currentTime
+        let duration = playProgress.duration
+        let timeProgress = currentTime / duration
+        updateProgressWaveform(timeProgress)
+    }
+    
+    func updateProgressWaveform(_ progress: Double) {
+        let fullRect = progressView.bounds
+        let newWidth = Double(fullRect.size.width) * progress
+        let maskLayer = CAShapeLayer()
+        let maskRect = CGRect(x: 0.0, y: 0.0, width: newWidth, height: Double(fullRect.size.height))
+        let path = CGPath(rect: maskRect, transform: nil)
+        maskLayer.path = path
+        progressView.layer.mask = maskLayer
+    }
+    
+    func changeButtonImage() {
+        if audioPlayHelper.isPlaying {
+            DispatchQueue.main.async {
+                self.playButton.isHidden = false
+                self.playButton.setImage(self.playButtonImagePause, for: .normal)
+            }
+        }
+        if !audioPlayHelper.isPlaying {
+            DispatchQueue.main.async {
+                self.playButton.isHidden = false
+                self.playButton.setImage(self.playButtonImagePlay, for: .normal)
+            }
+        }
+    }
+    
+}
+
